requirements.txt
telethon>=1.36.0
pysocks>=1.7.1
pyyaml>=6.0
aiosqlite>=0.19.0
python-socks[asyncio]>=2.4.0
rich>=13.0.0
